The overlap needed hand-resolving: 424bfbcf fixed NesEverdriveFifo.hpp's Windows pid include, and
de4f76c5 had moved that helper out to util/ProcessId.hpp - so the fix had to land in its new home
rather than be dropped with the old one.

util/ProcessId.hpp now carries the <process.h> / ::_getpid() form verbatim, and processAlive (added
after that, for the scratch sweep) moves to a new ProcessId.cpp: it needs OpenProcess, so it needs
<windows.h>, and this header is included by NesEverdriveFifo.hpp - i.e. by the whole NES core. One TU
pays for <windows.h> instead of all of them, which is the same argument 424bfbcf made.

Verified on the merged tree: 145 TS, 117 native, 29 UI, 8 plugin binaries, test:n8, and a linux build
clean of new warnings.

+++ packages/native/sdl/main.cpp
@@ -1499,7 +1499,7 @@ void applyWindowSize(AppState& a, std::uint32_t w, std::uint32_t h, bool fromWm)
     const std::uint32_t stride = lv_draw_buf_width_to_stride(w, cf);
     a.drawBuf.assign(static_cast<std::size_t>(stride) * h, 0);
     lv_display_set_resolution(a.display, w, h);
-    lv_display_set_buffers(a.display, a.drawBuf.data(), nullptr, a.drawBuf.size(), LV_DISPLAY_RENDER_MODE_DIRECT);
+    lv_display_set_buffers(a.display, a.drawBuf.data(), nullptr, static_cast<std::uint32_t>(a.drawBuf.size()), LV_DISPLAY_RENDER_MODE_DIRECT);
 
     // SDL streaming texture: recreate at the new size (can't be resized in place).
     if (a.texture) SDL_DestroyTexture(a.texture);
